requirements.txt
fastapi
ujson
aiofiles
uvicorn
mako
pyyaml
sqlalchemy
pymysql==0.9.2
aiomysql
databases[mysql]
python-multipart
alembic
pyjwt
passlib[bcrypt]
aiofiles
aioredis
pangu
mistune
FastAPI-Mako==0.6.0
